refactor(core): restructure project into core, zig, zls, pkg modules
- move util.lua, config.lua, ffi.lua, cmd.lua to core/ as core_util.lua, core_config.lua, core_ffi.lua, core_cmd.lua
- move module/zls.lua to zls/init.lua and update to use core layer
- move module/zig.lua to zig/init.lua and update to use core layer
- move module/pkg.lua to pkg/init.lua and update to use core and zig/zls
- move module/init.lua to info.lua for info panel logic
- update all require paths to use new structure
- update plugin/zig-lamp.lua to load new modules
- update ftplugin/zig.lua and health.lua to use core modules
- add zig-lamp-refactor-plan.md documenting refactor stages and goals
- remove old module/pkg.lua, module/init.lua, module/zig.lua, module/zls.lua
- preserve and adapt all original interfaces and commands
