+++ lua/zig-lamp/core/core_ffi.lua
@@ -1,10 +1,14 @@
+-- core_ffi.lua
+-- ffi 相关，原 ffi.lua 内容迁移于此
+
+local vim = vim
 local ffi = require("ffi")
 local path = require("plenary.path")
-local util = require("zig-lamp.util")
+local util = require("zig-lamp.core.core_util")
 
 local M = {}
 
-ffi.cdef([[
+ffi.cdef([[ 
     bool check_shasum(const char* file_path, const char* shasum);
     const char* get_build_zon_info(const char* file_path);
     void free_build_zon_info();
@@ -12,26 +16,39 @@ ffi.cdef([[
     void free_fmt_zon();
 ]])
 
--- stylua: ignore
-local plugin_path = vim.fs.normalize(string.sub(debug.getinfo(1).source, 2, #"/ffi.lua" * -1) .. "../../")
-
-local library_path = vim.fs.normalize((function()
-    if package.config:sub(1, 1) == "\\" then
-        return vim.fs.joinpath(plugin_path, "zig-out/bin/zig-lamp.dll")
-    else
-        if vim.fn.has("macunix") == 1 then
-            return vim.fs.joinpath(plugin_path, "zig-out/lib/libzig-lamp.dylib")
+local function find_plugin_path()
+    local dir = vim.fs.dirname(debug.getinfo(1).source:sub(2))
+    while dir do
+        -- 检查是否存在 lua/zig-lamp/core/core_ffi.lua 结构
+        local candidate = vim.fs.joinpath(dir, "lua", "zig-lamp", "core", "core_ffi.lua")
+        if vim.loop.fs_stat(candidate) then
+            return dir
         end
-        return vim.fs.joinpath(plugin_path, "zig-out/lib/libzig-lamp.so")
+        local parent = vim.fs.dirname(dir)
+        if parent == dir then break end
+        dir = parent
     end
-end)())
+    return nil
+end
+local plugin_path = vim.fs.normalize(find_plugin_path() or ".")
 
---- @type ffi.namespace*|nil|true
+local library_path = (function()
+    if package.config:sub(1, 1) == "\\" then
+        -- Windows
+        return vim.fs.normalize(vim.fs.joinpath(plugin_path, "zig-out", "bin", "zig-lamp.dll"))
+    else
+        -- Unix
+        if vim.fn.has("macunix") == 1 then
+            return vim.fs.normalize(vim.fs.joinpath(plugin_path, "zig-out", "lib", "libzig-lamp.dylib"))
+        else
+            return vim.fs.normalize(vim.fs.joinpath(plugin_path, "zig-out", "lib", "libzig-lamp.so"))
+        end
+    end
+end)()
+
 local _zig_lamp = nil
 
---- @return ffi.namespace*|nil
 function M.get_lamp()
-    -- when true, zig_lamp is not found
     if _zig_lamp == true then
         return nil
     end
@@ -47,8 +64,6 @@ function M.get_lamp()
     return nil
 end
 
--- whether zig-lamp is loaded
---- @return boolean
 function M.is_loaded()
     if _zig_lamp == nil or _zig_lamp == true then
         return false
@@ -56,7 +71,6 @@ function M.is_loaded()
     return true
 end
 
--- if zig-lamp is true, load it
 function M.lazy_load()
     if _zig_lamp ~= true then
         return
@@ -67,15 +81,10 @@ function M.lazy_load()
     end
 end
 
---- @return string
 function M.get_plugin_path()
     return plugin_path
 end
 
--- check sha256 digest
---- @param file_path string
---- @param shasum string
---- @return boolean
 function M.check_shasum(file_path, shasum)
     local zig_lamp = M.get_lamp()
     if not zig_lamp then
@@ -86,33 +95,15 @@ function M.check_shasum(file_path, shasum)
     return zig_lamp.check_shasum(file_path, shasum)
 end
 
---- @class ZigDependency
---- @field url? string
---- @field hash? string
---- @field path? string
---- @field lazy? boolean
----
---- @class ZigBuildZon
---- @field name string
---- @field version string
---- @field fingerprint string
---- @field minimum_zig_version string|nil
---- @field dependencies { [string] : ZigDependency }
---- @field paths string[]
-
--- get build.zig.zon info
--- this will parse as a table
---- @param file_path string
---- @return ZigBuildZon|nil
 function M.get_build_zon_info(file_path)
     local zig_lamp = M.get_lamp()
-    -- stylua: ignore
-    if not zig_lamp then return nil end
-
+    if not zig_lamp then
+        return nil
+    end
     local _p = path:new(file_path)
-    -- stylua: ignore
-    if not _p:exists() then return nil end
-
+    if not _p:exists() then
+        return nil
+    end
     local res = ffi.string(zig_lamp.get_build_zon_info(file_path))
     if res == "" then
         return nil
@@ -122,21 +113,19 @@ function M.get_build_zon_info(file_path)
     return _tmp
 end
 
---- @deprecated not use this
 function M.free_build_zon_info()
     local zig_lamp = M.get_lamp()
-    -- stylua: ignore
-    if not zig_lamp then return end
+    if not zig_lamp then
+        return
+    end
     zig_lamp.free_build_zon_info()
 end
 
--- format zon code
---- @param source_code string
---- @return string|nil
 function M.fmt_zon(source_code)
     local zig_lamp = M.get_lamp()
-    -- stylua: ignore
-    if not zig_lamp then return nil end
+    if not zig_lamp then
+        return nil
+    end
     local res = ffi.string(zig_lamp.fmt_zon(source_code))
     if res == "" then
         return nil
@@ -145,11 +134,11 @@ function M.fmt_zon(source_code)
     return res
 end
 
---- @deprecated not use this
 function M.free_fmt_zon()
     local zig_lamp = M.get_lamp()
-    -- stylua: ignore
-    if not zig_lamp then return end
+    if not zig_lamp then
+        return
+    end
     zig_lamp.free_fmt_zon()
 end
