+++ lua/zig-lamp/zig/init.lua
@@ -1,14 +1,13 @@
-local job = require("plenary.job")
+-- zig/init.lua
+-- 迁移自 module/zig.lua，接口与原始保持一致
 
+local job = require("plenary.job")
 local M = {}
 
--- get zig version
 --- @return string|nil
 function M.version()
-    --- @diagnostic disable-next-line: missing-fields
     local _tmp = job:new({ command = "zig", args = { "version" } })
     local _result, _ = _tmp:sync()
-
     if _result and #_result > 0 then
         return _result[1]
     end
