+++ plugin/zig-lamp.lua
@@ -1,5 +1,5 @@
-require("zig-lamp.cmd").init_command()
-require("zig-lamp.module").setup()
-require("zig-lamp.module.pkg").setup()
-require("zig-lamp.module.zig").setup()
-require("zig-lamp.module.zls").setup()
+require("zig-lamp.core.core_cmd").init_command()
+require("zig-lamp.info").setup()
+require("zig-lamp.pkg").setup()
+require("zig-lamp.zig").setup()
+require("zig-lamp.zls").setup()
